+++ DeBroglie/Constraints/ITileConstraint.cs
@@ -10,13 +10,13 @@
         /// </summary>
         /// <param name="propagator">The propagator to constraint</param>
         /// <returns>Contradiction if something is wrong, Undecided if generation should continue</returns>
-        Resolution Init(TilePropagator propagator);
+        void Init(TilePropagator propagator);
 
         /// <summary>
         /// Called frequently during generation to help maintain the constraint.
         /// </summary>
         /// <param name="propagator">The propagator to constraint</param>
         /// <returns>Contradiction if something is wrong, Undecided if generation should continue</returns>
-        Resolution Check(TilePropagator propagator);
+        void Check(TilePropagator propagator);
     }
 }