+++ DeBroglie/Constraints/PathConstraint.cs
@@ -31,14 +31,13 @@ namespace DeBroglie.Constraints
             this.EndPoints = endPoints;
         }
 
-        public Resolution Init(TilePropagator propagator)
+        public void Init(TilePropagator propagator)
         {
             tileSet = propagator.CreateTileSet(Tiles);
             graph = PathConstraintUtils.CreateGraph(propagator.Topology);
-            return Resolution.Undecided;
         }
 
-        public Resolution Check(TilePropagator propagator)
+        public void Check(TilePropagator propagator)
         {
             var topology = propagator.Topology;
             var indices = topology.Width * topology.Height * topology.Depth;
@@ -63,7 +62,7 @@ namespace DeBroglie.Constraints
             {
                 relevant = new bool[indices];
                 if (EndPoints.Length == 0)
-                    return Resolution.Undecided;
+                    return;
                 foreach (var endPoint in EndPoints)
                 {
                     var index = topology.GetIndex(endPoint.X, endPoint.Y, endPoint.Z);
@@ -76,7 +75,8 @@ namespace DeBroglie.Constraints
 
             if (isArticulation == null)
             {
-                return Resolution.Contradiction;
+                propagator.SetContradiction();
+                return;
             }
 
 
@@ -90,8 +90,6 @@ namespace DeBroglie.Constraints
                     propagator.Select(x, y, z, tileSet);
                 }
             }
-
-            return Resolution.Undecided;
         }
     }
 }